# docs/api-openapi.yaml
openapi: 3.1.0
info:
  title: SoloTasks API
  version: 1.0.0
servers:
  - url: https://api.example.com
  - url: http://localhost:8000
paths:
...
  /v1/sync/push:
    post:
      tags: [sync]
      summary: Push queued operations
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncPushInput'
      responses:
        '200':
          description: Push result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncPushResponse'
  /v1/sync/pull:
    get:
      tags: [sync]
      summary: Pull changes by cursor
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: cursor
          schema: { type: integer, default: 0 }
        - in: query
          name: deviceId
          required: true
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 500 }
      responses:
        '200':
          description: Pulled changes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncPullResponse'
  /v1/tasks:
    get:
      tags: [tasks]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Task list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Task' }
    post:
      tags: [tasks]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskCreate' }
      responses:
        '201':
          description: Task created
  /v1/tasks/{taskId}:
    put:
      tags: [tasks]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: taskId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskUpdate' }
      responses:
        '200':
          description: Task updated
    delete:
      tags: [tasks]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: taskId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Task deleted }
  /v1/tasks/series:
    post:
      tags: [tasks]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskSeriesCreate' }
      responses:
        '201': { description: Series created }
  /v1/tasks/{taskId}/subtasks:
    get:
      tags: [subtasks]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: taskId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Subtask list }
    post:
      tags: [subtasks]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: taskId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SubtaskCreate' }
      responses:
        '201': { description: Subtask created }
  /v1/tags:
    get:
      tags: [tags]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Tag list }
    post:
      tags: [tags]
      security: [{ bearerAuth: [] }]
      responses:
        '201': { description: Tag created }
  /v1/reminders:
    get:
      tags: [reminders]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Reminder list }
    post:
      tags: [reminders]
      security: [{ bearerAuth: [] }]
      responses:
        '201': { description: Reminder created }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Task:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        status: { type: string, enum: [todo, in_progress, blocked, done] }
        date_local: { type: string, format: date }
        start_min: { type: integer }
        end_min: { type: integer, nullable: true }
        duration_min: { type: integer, nullable: true }
        series_id: { type: string, nullable: true }
        series_index: { type: integer, nullable: true }
        series_total: { type: integer, nullable: true }
    TaskCreate:
      type: object
      required: [title, dateLocal, startMin]
      properties:
        title: { type: string }
        description: { type: string }
        status: { type: string, enum: [todo, in_progress, blocked, done] }
        dateLocal: { type: string, format: date }
        startMin: { type: integer }
        endMin: { type: integer }
        durationMin: { type: integer }
    TaskUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        status: { type: string, enum: [todo, in_progress, blocked, done] }
        dateLocal: { type: string, format: date }
        startMin: { type: integer }
        endMin: { type: integer }
        durationMin: { type: integer }
    TaskSeriesCreate:
      type: object
      required: [baseTitle, startDate, startMin]
      properties:
        baseTitle: { type: string }
        description: { type: string }
        status: { type: string, enum: [todo, in_progress, blocked, done] }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        numberOfDays: { type: integer }
        startMin: { type: integer }
        endMin: { type: integer }
        durationMin: { type: integer }
    SubtaskCreate:
      type: object
      required: [title]
      properties:
        title: { type: string }
        isDone: { type: boolean }
        orderKey: { type: string }
        note: { type: string }
...
    SyncPushInput:
      type: object
      required: [deviceId, ops]
      properties:
        deviceId: { type: string }
        ops:
          type: array
          items:
            type: object
            properties:
              opId: { type: string }
              entity: { type: string }
              action: { type: string }
              entityId: { type: string }
              payload: { type: object }
              clientTs: { type: string, format: date-time }
    SyncPushResponse:
      type: object
      properties:
        serverTime: { type: string, format: date-time }
        appliedCount: { type: integer }
        rejectedOps:
          type: array
          items:
            type: object
            properties:
              opId: { type: string }
              reason: { type: string }
    SyncPullResponse:
      type: object
      properties:
        cursor: { type: integer }
        serverTime: { type: string, format: date-time }
        changes:
          type: array
          items:
            type: object
            properties:
              seq: { type: integer }
              entity: { type: string }
              action: { type: string }
              entityId: { type: string }
              record: { type: object }
              deletedMeta: { type: object }